+++ src/game-manager/model.js
