+++ src/index.js
@@ -1,8 +1,8 @@
-import UserInputModel from "./model/userInput.js";
-import UserInputView from "./view/userInput.js";
+import UserInputModel from "./input-checker/model.js";
+import UserInputView from "./input-checker/view.js";
 
-import GameRuleModel from "./model/gameRules.js";
-import GameRuleView from "./view/gameRules.js";
+import GameRuleModel from "./game-manager/model.js";
+import GameRuleView from "./game-manager/view.js";
 
 import { RESULT_CODE } from "./constants.js";
