+++ src/input-checker/model.js
