+++ src/input-checker/view.js
